Move getMarkOptions to own file
