+++ src/getPlotOptions.ts
@@ -15,110 +15,9 @@ import type {
   Indexable,
   Sorts,
 } from "./types";
+import { borderOptions, defaultColors } from "./helpers";
 // Extend the MarkOptions to include all the stack options
 interface AllMarkOptions extends MarkOptions, StackOptions {}
-export const defaultColors = [
-  "rgba(255, 0, 184, 1)",
-  "rgba(0, 183, 255, 1)",
-  "rgba(255, 237, 0, 1)",
-  "rgba(0, 202, 99, 1)",
-  "rgba(255, 83, 0, 1)",
-];
-const borderOptions = {
-  backgroundColor: "hsla( 0 0% 100%)",
-  borderColor: "rgb(228, 229, 231)",
-};
-// Get options for a specific mark (e.g., the line or area marks)
-export function getMarkOptions(
-  currentColumns: string[] = [],
-  type: ChartType,
-  colTypes:
-    | {
-        [key: string]: BasicColumnType;
-      }
-    | undefined,
-  options: {
-    color?: string;
-    xLabel?: string;
-    yLabel?: string;
-    tip?: boolean;
-    xValue?: (d: Indexable, i: number) => string;
-    yValue?: (d: Indexable, i: number) => string;
-    markOptions?: AllMarkOptions;
-  }
-) {
-  const color = options.color || defaultColors[0];
-  const stroke = currentColumns.includes("series") ? "series" : color;
-  const fill = currentColumns.includes("series") ? "series" : color;
-  const fx = currentColumns.includes("fx") ? "fx" : undefined;
-  const tip =
-    options.tip !== false
-      ? {
-          tip: {
-            stroke: borderOptions.borderColor,
-            format: {
-              xCustom: true,
-              yCustom: true,
-              x: false,
-              y: false,
-              color: true,
-              fy: false,
-              fx: false,
-              z: false, // Hide the auto generated "series" for area charts
-            },
-          },
-        }
-      : {};
-  const ellipsis = "…";
-  function truncateLabel(label: string | undefined, length: number = 25) {
-    if (!label || label.length < length) return label;
-    return label.slice(0, length) + ellipsis;
-  }
-
-  const sort =
-    options.markOptions?.sort ?? (colTypes?.x !== "string" && type !== "barX")
-      ? { sort: (d: any) => d.x }
-      : {};
-
-  return {
-    // Create custom labels for x and y (important if the labels are custom but hidden!)
-    channels: {
-      xCustom: {
-        label: truncateLabel(options.xLabel),
-        // TODO: good for grouped bar charts, not good for other fx
-        value:
-          typeof options.xValue === "function"
-            ? options.xValue
-            : currentColumns.includes("fx")
-            ? "fx"
-            : "x",
-      },
-      yCustom: {
-        label: truncateLabel(options.yLabel),
-        value: typeof options.yValue === "function" ? options.yValue : "y",
-      },
-    },
-    ...tip,
-    ...(type === "line" ? { stroke } : { fill }),
-    ...(currentColumns.includes("x") ? { x: `x` } : {}),
-    ...(sort ? sort : {}),
-    ...(currentColumns.includes("fy") ? { fy: "fy" } : {}),
-    ...(type === "dot" && currentColumns.includes("r") ? { r: "r" } : {}),
-    ...(type === "text" && currentColumns.includes("text")
-      ? { text: "text" }
-      : {}),
-    ...(fx ? { fx: `fx` } : {}),
-    ...(currentColumns.includes("y") ? { y: `y` } : {}),
-    ...(options.markOptions ? { ...options.markOptions } : {}),
-    ...(currentColumns.includes("series")
-      ? {
-          [type === "line" || type.startsWith("rule") || type.startsWith("tick")
-            ? "stroke"
-            : "fill"]: `series`,
-        }
-      : {}),
-  } satisfies MarkOptions;
-}
 
 // Identify the data currently in the dataset
 export function getDataOrder(data: ChartData | undefined, column: string) {
